+++ src/common/covid_hosp/utils.py
